backend/requirements.txt
Flask==2.2.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
werkzeug==2.2.2
PyJWT==2.8.0
requests==2.31.0
reportlab==4.0.8
svglib==1.5.1
Pillow==10.4.0
